third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-cascade-trick-2-expected.txt

This is a testharness.js-based test.
FAIL Custom property sets can be extended with the (self-referential) cascade trick assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
